Moved decorator definition
